+++ inlang/source-code/cli/src/commands/machine/translate.ts
@@ -9,7 +9,6 @@ import { projectOption } from "../../utilities/globalFlags.js"
 import progessBar from "cli-progress"
 import plimit from "p-limit"
 import type { Result } from "@inlang/result"
-import { toV1Message, fromV1Message } from "@inlang/sdk/v2"
 
 const rpcTranslateAction = process.env.MOCK_TRANSLATE_LOCAL
 	? mockMachineTranslateMessage
@@ -100,9 +99,11 @@ export async function translateCommandAction(args: { project: InlangProject }) {
 			return
 		}
 
-		const allMessages = v2Persistence
-			? (await args.project.store!.messageBundles.getAll()).map(toV1Message)
-			: args.project.query.messages.getAll()
+		if (v2Persistence) {
+			throw new Error("not implemented vor v2 Persistence")
+			// allMessages = (await args.project.store!.messageBundles.getAll()).map(toV1Message)
+		}
+		const allMessages = args.project.query.messages.getAll()
 
 		const filteredMessages = allMessages.filter((message) =>
 			hasMissingTranslations(message, sourceLanguageTag, targetLanguageTags)
@@ -140,7 +141,8 @@ export async function translateCommandAction(args: { project: InlangProject }) {
 				translatedMessage?.variants.length > toBeTranslatedMessage.variants.length
 			) {
 				if (v2Persistence) {
-					await args.project.store!.messageBundles.set({ data: fromV1Message(translatedMessage) })
+					throw new Error("not implemented vor v2 Persistence")
+					// await args.project.store!.messageBundles.set({ data: fromV1Message(translatedMessage) })
 				} else {
 					args.project.query.messages.update({
 						where: { id: translatedMessage.id },