+++ inlang/source-code/sdk/slotfile-webapp/src/messageBundleListReact.tsx
@@ -1,11 +1,9 @@
 import React, { useEffect, useState } from "react"
 import { storage } from "./storage/db-messagebundle.js"
-import { MessageBundleRxType } from "./storage/schema-messagebundle.js"
 import { createComponent } from "@lit/react"
 import { InlangMessageBundle } from "@inlang/message-bundle-component"
 import { mockSetting } from "./mock/settings.js"
-import { MessageBundle } from "../../src/v2/types.js"
-
+import { MessageBundle } from "../../src/v2/types/message-bundle.js"
 
 export const MessageBundleComponent = createComponent({
 	tagName: "inlang-message-bundle",
@@ -17,15 +15,15 @@ export const MessageBundleComponent = createComponent({
 })
 
 export function MessageBundleList() {
-	const [bundles, setBundles] = useState([] as MessageBundleRxType[])
-	const [db, setDb] = useState<any>()
+	const [bundles, setBundles] = useState([] as MessageBundle[])
+	const [messageBundleCollection, setMessageBundleCollection] = useState<any>()
 
 	useEffect(() => {
 		let query = undefined as any
 		;(async () => {
-			const db$ = (await storage).database
-			setDb(db$)
-			query = db$.messageBundles
+			const mc = (await storage).inlangProject.messageBundleCollection
+			setMessageBundleCollection(mc)
+			query = mc
 				.find()
 				//.sort({ updatedAt: "desc" })
 				.$.subscribe((bundles) => {
@@ -39,7 +37,7 @@ export function MessageBundleList() {
 
 	const onBundleChange = (messageBundle: { detail: { argument: MessageBundle } }) => {
 		// eslint-disable-next-line no-console
-		db?.messageBundles.upsert(messageBundle.detail.argument as any)
+		messageBundleCollection?.upsert(messageBundle.detail.argument)
 	}
 
 	return (