+++ inlang/source-code/sdk/src/v2/helper.ts
@@ -1,14 +1,6 @@
 import { randomHumanId } from "../storage/human-id/human-readable-id.js"
 import { randomId } from "./randomId.js"
-import {
-	LanguageTag,
-	MessageBundle,
-	MessageBundleWithSlots,
-	Message,
-	MessageSlot,
-	Text,
-	Variant,
-} from "./types.js"
+import { LanguageTag, MessageBundle, Message, Text, Variant } from "./types/index.js"
 
 /**
  * create v2 MessageBundle with a random human ID
@@ -70,48 +62,4 @@ export function toTextElement(text: string): Text {
 		type: "text",
 		value: text,
 	}
-}
-
-// ****************************
-// WIP experimental persistence
-// ****************************
-
-/**
- * create MessageSlot for a locale (only used for persistence)
- */
-export function createMessageSlot(locale: LanguageTag): MessageSlot {
-	return {
-		locale,
-		slot: true,
-	}
-}
-
-/**
- * return structuredClone with message slots for all locales not yet present
- */
-export function addSlots(messageBundle: MessageBundle, locales: string[]): MessageBundleWithSlots {
-	const bundle = structuredClone(messageBundle) as MessageBundleWithSlots
-	bundle.messages = locales.map((locale) => {
-		return bundle.messages.find((message) => message.locale === locale) ?? createMessageSlot(locale)
-	})
-	return bundle
-}
-
-/**
- * remove empty message slots without first creating a structured clone
- */
-export function removeSlots(messageBundle: MessageBundleWithSlots) {
-	messageBundle.messages = messageBundle.messages.filter((message) => !("slot" in message))
-	return messageBundle as MessageBundle
-}
-
-/**
- * Add newlines between bundles and messages to avoid merge conflicts
- */
-export function injectJSONNewlines(json: string): string {
-	return json
-		.replace(/\{"id":"/g, '\n\n\n\n{"id":"')
-		.replace(/"messages":\[\{"locale":"/g, '"messages":[\n\n\n\n{"locale":"')
-		.replace(/\}\]\}\]\},\{"locale":"/g, '}]}]},\n\n\n\n{"locale":"')
-		.replace(/"slot":true\},\{"locale":/g, '"slot":true},\n\n\n\n{"locale":')
 }