+++ inlang/source-code/cross-sell/cross-sell-ninja/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @inlang/cross-sell-ninja
 
+## 0.0.34
+
+### Patch Changes
+
+- Updated dependencies [2fc5feb]
+  - @inlang/sdk@0.36.2
+
+## 0.0.33
+
+### Patch Changes
+
+- Updated dependencies [1077e06]
+  - @inlang/sdk@0.36.1
+
+## 0.0.32
+
+### Patch Changes
+
+- Updated dependencies [8ec7b34]
+- Updated dependencies [05f9282]
+  - @inlang/sdk@0.36.0
+
 ## 0.0.31
 
 ### Patch Changes