+++ inlang/source-code/doc-layout-component/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @inlang/settings-component
 
+## 0.0.22
+
+### Patch Changes
+
+- Updated dependencies [2fc5feb]
+  - @inlang/sdk@0.36.2
+
+## 0.0.21
+
+### Patch Changes
+
+- Updated dependencies [1077e06]
+  - @inlang/sdk@0.36.1
+
+## 0.0.20
+
+### Patch Changes
+
+- Updated dependencies [8ec7b34]
+- Updated dependencies [05f9282]
+  - @inlang/sdk@0.36.0
+
 ## 0.0.19
 
 ### Patch Changes